refactor(api,admin): standardize database ID types and enhance party forms
- Standardize database schema column and foreign key types across migrations:
  - election_groups.id -> SMALLINT (int16)
  - offices.id -> SMALLINT (int16)
  - parties.id -> SMALLINT (int16)
  - elections.id & polling_units.id -> INT (int32)
  - c_states.id -> SMALLINT (int16), lgas.id & wards.id -> INT (int32)
- Regenerate sqlc queries and models across all domains (agent earnings, election stats, party applications, assignments, voting, referrals, inec grabber)
- Align all Go services, handlers, routers, and background worker tasks to match strongly typed sqlc models
- Add date_founded support and refine UI layout/logo upload in admin PartyFormDialog, party-tile, and party server functions
- Update workspace dictionary settings in .vscode/settings.json

+++ apps/api/db/migrations/20260702115000_polling_unit_updates.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS polling_unit_updates (
   
   -- Denormalized references for fast dashboard filtering
   user_id BIGINT REFERENCES users(id) ON DELETE CASCADE NOT NULL,
-  election_group_id BIGINT REFERENCES election_groups(id) ON DELETE CASCADE NOT NULL,
+  election_group_id SMALLINT REFERENCES election_groups(id) ON DELETE CASCADE NOT NULL,
   polling_unit_id INTEGER REFERENCES polling_units(id) ON DELETE CASCADE NOT NULL,
   party_id SMALLINT REFERENCES parties(id) ON DELETE SET NULL,
   state_id SMALLINT REFERENCES c_states(id) ON DELETE SET NULL,
